#328 - changing tab title
gonna need reviews on this
