+++ sonar-project-frontend.properties
@@ -19,6 +19,7 @@ sonar.coverage.minimumCoverage=80
 
 # Exclude specific files from analysis
 sonar.exclusions=apps/frontend/src/app/layout.tsx
+sonar.exclusions=apps/frontend/src/app/client-layout.tsx
 
 # Encoding
 sonar.sourceEncoding=UTF-8
